Save full 'data' to local storage
Not just an array of 'checklists'. This allows us to save more fields
into the same blob.

The initial idea was to save in some indication of when you last changed
the structure and if it had been backed up but that is a bit too much
work or too messy so I'm going to come back later.